+++ python/paddle/fluid/contrib/slim/quantization/__init__.py
@@ -20,12 +20,15 @@ from . import quantization_strategy
 from .quantization_strategy import *
 from . import mkldnn_post_training_strategy
 from .mkldnn_post_training_strategy import *
-from . import quantization_mkldnn_pass
-from .quantization_mkldnn_pass import *
+from . import qat_int8_mkldnn_pass
+from .qat_int8_mkldnn_pass import *
+from . import qat2_int8_mkldnn_pass
+from .qat2_int8_mkldnn_pass import *
 from . import post_training_quantization
 from .post_training_quantization import *
 
 __all__ = quantization_pass.__all__ + quantization_strategy.__all__
 __all__ += mkldnn_post_training_strategy.__all__
-__all__ += quantization_mkldnn_pass.__all__
+__all__ += qat_int8_mkldnn_pass.__all__
+__all__ += qat2_int8_mkldnn_pass.__all__
 __all__ += post_training_quantization.__all__
